refactor(http): share HTTP/1 header+body writing between request/response
write_http_request and write_http_response carried byte-identical blocks for
(a) serializing the Http1HeaderMap to the wire and (b) collecting/writing the
body. Extract both into write_http1_header_map and write_http1_body helpers in
io/mod.rs and call them from each writer.

Behavior unchanged; io request/response tests green, clippy clean.
